- Fixed bug that would cause the AbundanceSumRaw to be Int.MinValue if it was too large - Upgrading to latest and greatest Math.NET (called Numerics instead of Iridium now) - Deleted some unused methods that I could not get to compile with the latest Math.NET

+++ FeatureFinder/Data/LCIMSMSFeature.cs
@@ -20,8 +20,8 @@ namespace FeatureFinder.Data
 		public double DriftTime { get; set; }
 		public int OriginalIndex { get; set; }
 		public int MaxMemberCount { get; set; }
-		public int AbundanceMaxRaw { get; set; }
-		public int AbundanceSumRaw { get; set; }
+		public double AbundanceMaxRaw { get; set; }
+		public double AbundanceSumRaw { get; set; }
 
 		public LCIMSMSFeature(byte charge)
 		{
